Upgrade rollup pipeline & fix minification
The closure compiler has been replaced by terser for the minification step, as
closure does not handle correctly variables (or in this case constants)
declared with the same name as some globals (in this case `document`). This is
a valid error for scripts, but in ES modules this kind of redeclarations do not
cause problems (at least for `document`).

For details see:
- https://github.com/ampproject/rollup-plugin-closure-compiler/issues/92
- https://github.com/google/closure-compiler/issues/3098
- https://github.com/google/closure-compiler/issues/2471

Also terser is way faster than closure, and produces a slightly smaller
minified script.